Make "Add related content" button look interactive
The button looked just like regular text and it wasn't obvious that it
was an interactive element.

So we're styling it like the "Hide" link (which should actually be a
button, by the way).

Note buttons by default don't have a `cursor: pointer` property because
they're usually styled in a way that makes it obvious they're
interactive elements. Since that isn't the case here, we're adding that
extra hint for mouse users.

Also note all these links/buttons will look like regular text to color-
blind users. And if they use a touchscreen, they won't be able to hover
on them to see the cursor change. We might need to change these elements
in the future.